requirements.txt
openai
sendgrid
python-dotenv
google-api-python-client
youtube-transcript-api
isodate

